+++ App.js
@@ -9,34 +9,22 @@ import {
   FlatList,
 } from "react-native";
 import GoalItem from "./components/goals/GoalItem";
+import GoalInput from "./components/goals/GoalInput";
 
 export default function App() {
-  const [inputGoal, setInputGoal] = useState("");
   const [courseGoals, setCourseGoals] = useState([]);
 
-  const inputChangeHandler = (enteredGoal) => {
-    setInputGoal(enteredGoal);
-  };
-
-  const addGoalHandler = () => {
-    console.log(inputGoal);
+  const addGoalHandler = (goalTitle) => {
+    console.log(goalTitle);
     setCourseGoals((courseGoals) => [
       ...courseGoals,
-      { id: Math.random().toString(), value: inputGoal },
+      { id: Math.random().toString(), value: goalTitle },
     ]);
   };
 
   return (
     <View style={styles.screen}>
-      <View style={styles.inputContainer}>
-        <TextInput
-          placeholder="Course Goal"
-          style={styles.input}
-          onChangeText={inputChangeHandler}
-          value={inputGoal}
-        />
-        <Button title="ADD" onPress={addGoalHandler} />
-      </View>
+      <GoalInput goalHandler={addGoalHandler} />
       {
         // this will give warning her as we dont have any key here and for that it expect certain format of data with key
       }
@@ -51,17 +39,6 @@ export default function App() {
 
 const styles = StyleSheet.create({
   screen: { padding: 50 },
-  inputContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-  },
-  input: {
-    borderColor: "black",
-    borderWidth: 1,
-    padding: 10,
-    width: "80%",
-  },
 });
 
 // add some styling to the view
